Speed up weather icon and bg pattern loading
This is aimed at users using a slow network connection. The idea is to
create image elements with weather icons when the Location class
is instantiated. Previously, weather icons for forecast tabs started
to be downloaded only after clicking any forecast tab. The solution is
not 100% solid, but still improves user experience.