Fix unavailable data that broke the entire app
Unfortunately, the Air Quality Index for forecast data is not available
on the free Weather API plan. The Pricing tab did not say that but the
docs say so (although vaguely). If I'd known it would be unavailable, I
would not have added it to the app. Solved it by removing both Air
Quality Index and UV Index from forecast tabs.